[mod#177] Kid 부모 탈퇴 여부 조회 추가 및 MyPage 패키지 구조 리팩토링
- Kid 부모 탈퇴 상태 조회(getParentWithdrawalStatus) 관련 data 레이어 추가
- Parent/Kid MyPage 관련 패키지 경로 및 디렉토리 구조 정리
- SseRepository 내 TokenManager를 UserInfoManager로 교체하여 유저 권한 확인
- Kid MyPage 관련 DI 모듈 및 서비스 코드 정리
