+++ app/src/main/java/com/kiero/data/kid/mypage/repository/KidMyPageRepository.kt
@@ -1,0 +1,7 @@
+package com.kiero.data.kid.mypage.repository
+
+import com.kiero.data.kid.mypage.model.KidTermsModel
+
+interface KidMyPageRepository {
+    suspend fun getKidTermsExternalLink(): Result<List<KidTermsModel>>
+}
