+++ app/src/main/java/com/kiero/data/kid/user/remote/datasource/KidUserDataSource.kt
@@ -1,0 +1,8 @@
+package com.kiero.data.kid.user.remote.datasource
+
+import com.kiero.core.network.model.BaseResponse
+import com.kiero.data.kid.user.remote.dto.response.KidParentWithdrawalStatusResponseDto
+
+interface KidUserDataSource {
+    suspend fun getParentWithdrawalStatus(): BaseResponse<KidParentWithdrawalStatusResponseDto>
+}
