+++ app/src/main/java/com/kiero/data/kid/user/repository/KidUserRepository.kt
@@ -1,0 +1,7 @@
+package com.kiero.data.kid.user.repository
+
+import com.kiero.data.kid.user.model.KidParentWithdrawalStatusModel
+
+interface KidUserRepository {
+    suspend fun getParentWithdrawalStatus(): Result<KidParentWithdrawalStatusModel>
+}
