--- app/src/main/java/com/kiero/data/parent/mypage/kid/remote/datasource/KidMyPageDataSource.kt
@@ -1,8 +1,0 @@
-package com.kiero.data.parent.mypage.kid.remote.datasource
-
-import com.kiero.core.network.model.BaseResponse
-import com.kiero.data.parent.mypage.kid.remote.dto.response.KidTermResponseDto
-
-interface KidMyPageDataSource {
-    suspend fun getKidTermsExternalLink(): BaseResponse<List<KidTermResponseDto>>
-}
