--- app/src/main/java/com/kiero/data/parent/mypage/kid/repository/KidMyPageRepository.kt
@@ -1,7 +1,0 @@
-package com.kiero.data.parent.mypage.kid.repository
-
-import com.kiero.data.parent.mypage.kid.model.KidTermsModel
-
-interface KidMyPageRepository {
-    suspend fun getKidTermsExternalLink(): Result<List<KidTermsModel>>
-}
