--- app/src/main/java/com/kiero/data/parent/mypage/parent/model/ParentMyProfileModel.kt
@@ -1,17 +1,0 @@
-package com.kiero.data.parent.mypage.parent.model
-
-import com.kiero.data.parent.mypage.parent.remote.dto.response.parent.ParentMyProfileResponseDto
-
-data class ParentMyProfileModel(
-    val image: String?,
-    val name: String,
-    val hasPendingChildSession: Boolean,
-    val pushNotificationEnabled: Boolean
-)
-
-fun ParentMyProfileResponseDto.toModel() = ParentMyProfileModel(
-    image = image,
-    name = name,
-    hasPendingChildSession = hasPendingChildSession,
-    pushNotificationEnabled = pushNotificationEnabled
-)
