--- app/src/main/java/com/kiero/data/parent/mypage/parent/model/ParentTermsModel.kt
@@ -1,13 +1,0 @@
-package com.kiero.data.parent.mypage.parent.model
-
-import com.kiero.data.parent.mypage.parent.remote.dto.response.parent.ParentTermResponseDto
-
-data class ParentTermsModel(
-    val linkType: String,
-    val link: String
-)
-
-fun ParentTermResponseDto.toModel() = ParentTermsModel(
-    linkType = linkType,
-    link = link
-)
