--- app/src/main/java/com/kiero/data/parent/mypage/parent/remote/dto/response/parent/ParentTermResponseDto.kt
@@ -1,12 +1,0 @@
-package com.kiero.data.parent.mypage.parent.remote.dto.response.parent
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ParentTermResponseDto(
-    @SerialName("linkType")
-    val linkType: String,
-    @SerialName("link")
-    val link: String
-)
