+++ app/src/main/java/com/kiero/data/parent/mypage/remote/dto/response/parent/ParentMyProfileResponseDto.kt
@@ -1,0 +1,16 @@
+package com.kiero.data.parent.mypage.remote.dto.response.parent
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ParentMyProfileResponseDto(
+    @SerialName("image")
+    val image: String?,
+    @SerialName("name")
+    val name: String,
+    @SerialName("hasPendingChildSession")
+    val hasPendingChildSession: Boolean,
+    @SerialName("pushNotificationEnabled")
+    val pushNotificationEnabled: Boolean
+)
