+++ app/src/main/java/com/kiero/data/parent/mypage/remote/dto/response/parent/ParentTermResponseDto.kt
@@ -1,0 +1,12 @@
+package com.kiero.data.parent.mypage.remote.dto.response.parent
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ParentTermResponseDto(
+    @SerialName("linkType")
+    val linkType: String,
+    @SerialName("link")
+    val link: String
+)
